move entry point to event.rs
