+++ relay-event-normalization/src/lib.rs
@@ -7,14 +7,12 @@
 )]
 
 use std::collections::BTreeSet;
-use std::ops::Range;
 use std::sync::Arc;
 
 use chrono::{DateTime, Utc};
-use relay_common::time::UnixTimestamp;
 use relay_event_schema::processor::{ProcessingResult, ProcessingState, Processor};
 use relay_event_schema::protocol::{Event, IpAddr, SpanAttribute};
-use relay_protocol::{Annotated, Meta};
+use relay_protocol::Meta;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use uuid::Uuid;
@@ -37,6 +35,7 @@ mod transactions;
 mod trimming;
 
 pub mod replay;
+pub use event::{normalize_event, NormalizationConfig};
 pub use normalize::breakdowns::*;
 pub use normalize::*;
 pub use transactions::*;
@@ -213,153 +212,4 @@ impl<'a> Processor for StoreProcessor<'a> {
 
         Ok(())
     }
-}
-
-/// Configuration for [`normalize_event`].
-#[derive(Clone, Debug)]
-pub struct NormalizationConfig<'a> {
-    /// The IP address of the SDK that sent the event.
-    ///
-    /// When `{{auto}}` is specified and there is no other IP address in the payload, such as in the
-    /// `request` context, this IP address gets added to the `user` context.
-    pub client_ip: Option<&'a IpAddr>,
-
-    /// The user-agent and client hints obtained from the submission request headers.
-    ///
-    /// Client hints are the preferred way to infer device, operating system, and browser
-    /// information should the event payload contain no such data. If no client hints are present,
-    /// normalization falls back to the user agent.
-    pub user_agent: RawUserAgentInfo<&'a str>,
-
-    /// The time at which the event was received in this Relay.
-    ///
-    /// This timestamp is persisted into the event payload.
-    pub received_at: Option<DateTime<Utc>>,
-
-    /// The maximum amount of seconds an event can be dated in the past.
-    ///
-    /// If the event's timestamp is older, the received timestamp is assumed.
-    pub max_secs_in_past: Option<i64>,
-
-    /// The maximum amount of seconds an event can be predated into the future.
-    ///
-    /// If the event's timestamp lies further into the future, the received timestamp is assumed.
-    pub max_secs_in_future: Option<i64>,
-
-    /// Timestamp range in which a transaction must end.
-    ///
-    /// Transactions that finish outside of this range are considered invalid.
-    /// This check is skipped if no range is provided.
-    pub transaction_range: Option<Range<UnixTimestamp>>,
-
-    /// The maximum length for names of custom measurements.
-    ///
-    /// Measurements with longer names are removed from the transaction event and replaced with a
-    /// metadata entry.
-    pub max_name_and_unit_len: Option<usize>,
-
-    /// Configuration for measurement normalization in transaction events.
-    ///
-    /// Has an optional [`crate::MeasurementsConfig`] from both the project and the global level.
-    /// If at least one is provided, then normalization will truncate custom measurements
-    /// and add units of known built-in measurements.
-    pub measurements: Option<DynamicMeasurementsConfig<'a>>,
-
-    /// Emit breakdowns based on given configuration.
-    pub breakdowns_config: Option<&'a BreakdownsConfig>,
-
-    /// When `Some(true)`, context information is extracted from the user agent.
-    pub normalize_user_agent: Option<bool>,
-
-    /// Configuration to apply to transaction names, especially around sanitizing.
-    pub transaction_name_config: TransactionNameConfig<'a>,
-
-    /// When `Some(true)`, it is assumed that the event has been normalized before.
-    ///
-    /// This disables certain normalizations, especially all that are not idempotent. The
-    /// renormalize mode is intended for the use in the processing pipeline, so an event modified
-    /// during ingestion can be validated against the schema and large data can be trimmed. However,
-    /// advanced normalizations such as inferring contexts or clock drift correction are disabled.
-    ///
-    /// `None` equals to `false`.
-    pub is_renormalize: bool,
-
-    /// When `true`, infers the device class from CPU and model.
-    pub device_class_synthesis_config: bool,
-
-    /// When `true`, extracts tags from event and spans and materializes them into `span.data`.
-    pub enrich_spans: bool,
-
-    /// When `true`, computes and materializes attributes in spans based on the given configuration.
-    pub normalize_spans: bool,
-
-    /// The maximum allowed size of tag values in bytes. Longer values will be cropped.
-    pub max_tag_value_length: usize, // TODO: move span related fields into separate config.
-
-    /// Configuration for replacing identifiers in the span description with placeholders.
-    ///
-    /// This is similar to `transaction_name_config`, but applies to span descriptions.
-    pub span_description_rules: Option<&'a Vec<SpanDescriptionRule>>,
-
-    /// Configuration for generating performance score measurements for web vitals
-    pub performance_score: Option<&'a PerformanceScoreConfig>,
-
-    /// An initialized GeoIP lookup.
-    pub geoip_lookup: Option<&'a GeoIpLookup>,
-
-    /// When `Some(true)`, individual parts of the event payload is trimmed to a maximum size.
-    ///
-    /// See the event schema for size declarations.
-    pub enable_trimming: bool,
-}
-
-impl<'a> Default for NormalizationConfig<'a> {
-    fn default() -> Self {
-        Self {
-            client_ip: Default::default(),
-            user_agent: Default::default(),
-            received_at: Default::default(),
-            max_secs_in_past: Default::default(),
-            max_secs_in_future: Default::default(),
-            transaction_range: Default::default(),
-            max_name_and_unit_len: Default::default(),
-            breakdowns_config: Default::default(),
-            normalize_user_agent: Default::default(),
-            transaction_name_config: Default::default(),
-            is_renormalize: Default::default(),
-            device_class_synthesis_config: Default::default(),
-            enrich_spans: Default::default(),
-            normalize_spans: Default::default(),
-            max_tag_value_length: usize::MAX,
-            span_description_rules: Default::default(),
-            performance_score: Default::default(),
-            geoip_lookup: Default::default(),
-            enable_trimming: false,
-            measurements: None,
-        }
-    }
-}
-
-/// Normalizes an event, rejecting it if necessary.
-///
-/// Normalization consists of applying a series of transformations on the event
-/// payload based on the given configuration.
-///
-/// The returned [`ProcessingResult`] indicates whether the passed event should
-/// be ingested or dropped.
-pub fn normalize_event(
-    event: &mut Annotated<Event>,
-    config: &NormalizationConfig,
-) -> ProcessingResult {
-    let Annotated(Some(ref mut event), ref mut meta) = event else {
-        return Ok(());
-    };
-
-    let is_renormalize = config.is_renormalize;
-
-    if !is_renormalize {
-        event::normalize(event, meta, config)?;
-    }
-
-    Ok(())
 }
